requirements.txt
dgl_cu110==0.6.1
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.0.2
scipy==1.7.3
torch==1.10.2+cu111
tqdm==4.66.1
